+++ src/components/CourierInboxPopupMenuVue.vue
@@ -1,30 +1,78 @@
 <template>
-  <div class="courier-inbox-vue">
-    <courier-inbox-popup-menu ref="inbox" v-bind="propsBinding"/>
+  <div class="courier-inbox-popup-menu-vue" :class="{ disabled: !userId }">
+    <courier-inbox-popup-menu ref="inbox" v-bind="propsBinding" />
   </div>
 </template>
 
 <script setup lang="ts">
-import { useTemplateRef } from 'vue';
+import { onMounted, ref, toRef, useTemplateRef, watch } from 'vue';
 import type { CourierInboxPopupMenuProps } from '../types';
-import { type CourierInboxPopupMenu, type CourierInboxListItemActionFactoryProps, type CourierInboxListItemFactoryProps } from '@trycourier/courier-ui-inbox';
+import type {
+  CourierInboxPopupMenu,
+  CourierInboxListItemActionFactoryProps,
+  CourierInboxListItemFactoryProps,
+  CourierInboxMenuButtonFactoryProps,
+  CourierInboxHeaderFactoryProps,
+  CourierInboxStateEmptyFactoryProps,
+  CourierInboxStateErrorFactoryProps,
+  CourierInboxStateLoadingFactoryProps,
+  CourierInboxPaginationItemFactoryProps
+} from '@trycourier/courier-ui-inbox';
 import { useKebabBinding } from '@/ts/useKebabBinding';
+import { useCourier } from '@/ts/useCourier';
+import { useInboxSlotRenderer } from './useInboxSlotRenderer';
+
+type SlotProps<T> = T & { ctx: CourierInboxPopupMenu };
+
+const props = withDefaults(defineProps<CourierInboxPopupMenuProps>(), { mode: 'light' });
+const propsBinding = useKebabBinding(props);
+
+const slots = defineSlots<{
+  button(props: SlotProps<CourierInboxMenuButtonFactoryProps>): any;
+  header(props: SlotProps<CourierInboxHeaderFactoryProps>): any;
+  item(props: SlotProps<CourierInboxListItemFactoryProps>): any;
+  empty(props: SlotProps<CourierInboxStateEmptyFactoryProps>): any;
+  error(props: SlotProps<CourierInboxStateErrorFactoryProps>): any;
+  loading(props: SlotProps<CourierInboxStateLoadingFactoryProps>): any;
+  pagination(props: SlotProps<CourierInboxPaginationItemFactoryProps>): any;
+}>();
+
+const { userId } = useCourier();
 
 const inbox = useTemplateRef<CourierInboxPopupMenu>('inbox');
 
-const props = withDefaults(defineProps<CourierInboxPopupMenuProps>(), {mode: "light"});
-
-const propsBinding = useKebabBinding(props);
+useInboxSlotRenderer(toRef(slots, 'button'), inbox, ibx => ibx.setMenuButton);
+useInboxSlotRenderer(toRef(slots, 'header'), inbox, ibx => ibx.setHeader);
+useInboxSlotRenderer(toRef(slots, 'item'), inbox, ibx => ibx.setListItem);
+useInboxSlotRenderer(toRef(slots, 'empty'), inbox, ibx => ibx.setEmptyState);
+useInboxSlotRenderer(toRef(slots, 'error'), inbox, ibx => ibx.setErrorState);
+useInboxSlotRenderer(toRef(slots, 'loading'), inbox, ibx => ibx.setLoadingState);
+useInboxSlotRenderer(toRef(slots, 'pagination'), inbox, ibx => ibx.setPaginationItem);
 
 const emits = defineEmits<{
-  (e: 'messageClicked', value: CourierInboxListItemFactoryProps): void;
-  (e: 'messageActionClicked', value: CourierInboxListItemActionFactoryProps): void;
-  (e: 'messageLongPressed', value: CourierInboxListItemFactoryProps): void;
+  (e: 'message:clicked', value: CourierInboxListItemFactoryProps): void;
+  (e: 'message:actionClicked', value: CourierInboxListItemActionFactoryProps): void;
+  (e: 'message:longPressed', value: CourierInboxListItemFactoryProps): void;
 }>();
 
+watch(inbox, ibx => {
+  if (!ibx) {
+    return;
+  }
 
-inbox.value?.onMessageClick((props) => emits('messageClicked', props))
-inbox.value?.onMessageActionClick((props) => emits('messageActionClicked', props))
-inbox.value?.onMessageLongPress((props) => emits('messageLongPressed', props))
-
+  ibx.onMessageClick(props => emits('message:clicked', props));
+  ibx.onMessageActionClick(props => emits('message:actionClicked', props));
+  ibx.onMessageLongPress(props => emits('message:longPressed', props));
+});
 </script>
+
+<style>
+.courier-inbox-popup-menu-vue.disabled {
+  opacity: 0.5;
+  pointer-events: none;
+}
+
+.courier-inbox-popup-menu-vue.disabled courier-inbox-popup-menu #unread-badge {
+  display: none !important;
+}
+</style>