+++ src/index.ts
@@ -2,7 +2,14 @@ import CourierInboxVue from './components/CourierInboxVue.vue';
 import CourierInboxPopupMenuVue from './components/CourierInboxPopupMenuVue.vue';
 import CourierToastVue from './components/CourierToastVue.vue';
 
-import { useCourier} from './ts/useCourier';
+import {
+  useCourier,
+  useCourierBrands,
+  useCourierLists,
+  useCourierPreferences,
+  useCourierTokens,
+  useCourierTracking
+} from './ts/useCourier';
 
 export * from './types';
 
@@ -11,4 +18,9 @@ export {
   CourierInboxPopupMenuVue,
   CourierToastVue,
   useCourier,
+  useCourierBrands,
+  useCourierLists,
+  useCourierPreferences,
+  useCourierTokens,
+  useCourierTracking
 };