+++ vite.config.ts
@@ -4,7 +4,7 @@ import vue from '@vitejs/plugin-vue';
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 import { fileURLToPath } from 'node:url';
-import vueDevTools from 'vite-plugin-vue-devtools'
+import vueDevTools from 'vite-plugin-vue-devtools';
 
 const fileName = {
   es: 'index.es.mjs',
@@ -25,7 +25,7 @@ export default defineConfig({
     dts({
       tsconfigPath: 'tsconfig.build.json'
     }),
-    vueDevTools(),
+    vueDevTools()
   ],
   build: {
     minify: true,